log refactor 3
