--- tools/logger_patch.py
@@ -1,45 +1,0 @@
-"""
-# Copyright (c) 2025  PaddlePaddle Authors. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License"
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-"""
-
-import logging
-from contextlib import contextmanager
-
-
-@contextmanager
-def intercept_paddle_loggers():
-    """Intercept and configure paddle loggers during import."""
-    _original = logging.getLogger
-
-    def _patched(name=None):
-        logger = _original(name)
-        if name and str(name).startswith("paddle"):
-            formatter = logging.Formatter(
-                "%(levelname)-8s %(asctime)s %(process)-5s %(filename)s[line:%(lineno)d] %(message)s"
-            )
-            logger.setLevel(logging.INFO)
-            for handler in logger.handlers[:]:
-                logger.removeHandler(handler)
-            stream_handler = logging.StreamHandler()
-            stream_handler.setFormatter(formatter)
-            logger.addHandler(stream_handler)
-            logger.propagate = False
-        return logger
-
-    logging.getLogger = _patched
-    try:
-        yield
-    finally:
-        logging.getLogger = _original
